Kmeans works again
Former-commit-id: 52aeb3b854bc3cae2f3e67e135771090590a4376
Former-commit-id: d51e8045f204ad6e750dcbb7d2c55b1b861e0ab1
Former-commit-id: 7e6e45b3f2c720fdb9dee62382ff1a37500e8048

+++ benchmarks/amdappsdk/bitonicsort/bitonicsort.go
@@ -73,7 +73,7 @@ func (b *Benchmark) initMem() {
 	b.gInputData = b.driver.AllocateMemory(b.context, uint64(b.Length*4))
 	b.driver.Distribute(
 		b.context,
-		uint64(b.gInputData), uint64(b.Length*4),
+		b.gInputData, uint64(b.Length*4),
 		b.gpusToUse)
 
 	b.driver.MemCopyH2D(b.context, b.gInputData, b.inputData)
